PP9-8579: content download dialog logic optimization
