+++ src/picturepark-sdk-v1-angular/projects/picturepark-sdk-v1-angular-ui/src/public_api.ts
@@ -3,9 +3,23 @@
  */
 export { PICTUREPARK_UI_CONFIGURATION, PictureparkUIConfigurationFactory } from './lib/configuration';
 
-export { ContentItemSelectionService } from './lib/shared-module/services/content-item-selection/content-item-selection.service';
-export { BasketService } from './lib/shared-module/services/basket/basket.service';
-export { TranslationService } from './lib/shared-module/services/translations/translation.service';
+// export { ContentItemSelectionService } from './lib/shared-module/services/content-item-selection/content-item-selection.service';
+// export { BasketService } from './lib/shared-module/services/basket/basket.service';
+// export { TranslationService } from './lib/shared-module/services/translations/translation.service';
+// export { LiquidRenderingService } from './lib/shared-module/services/liquid-rendering/liquid-rendering.service';
+// export { DownloadFallbackService } from './lib/shared-module/services/download-fallback/download-fallback.service';
+
+// MODULES
 export { PictureparkUiModule } from './lib/picturepark-ui.module';
-export { LiquidRenderingService } from './lib/shared-module/services/liquid-rendering/liquid-rendering.service';
-export { DownloadFallbackService } from './lib/shared-module/services/download-fallback/download-fallback.service';
+
+export { BasketModule } from './lib/features-module/basket/basket.module';
+export { ContentAggregationListModule } from './lib/features-module/content-aggregation-list/content-aggregation-list.module';
+export { ContentBrowserModule } from './lib/features-module/content-browser/content-browser.module';
+export { ChannelPickerModule } from './lib/features-module/channel-picker/channel-picker.module';
+export { DatePickerModule } from './lib/features-module/date-picker/date-picker.module';
+export { DialogModule } from './lib/features-module/dialog/dialog.module';
+export { LayerPanelsModule } from './lib/features-module/layer-panels/layer-panels.module';
+export { ListItemAggregationListModule } from './lib/features-module/list-item-aggregation-list/list-item-aggregation-list.module';
+export { NotificationModule } from './lib/features-module/notification/notification.module';
+export { OutputDownloadMenuModule } from './lib/features-module/output-download-menu/output-download-menu.module';
+export { SearchBoxModule } from './lib/features-module/search-box/search-box.module';
